+++ receiver/hostmetricsreceiver/internal/scraper/processesscraper/processes_scraper_unix.go
@@ -45,7 +45,7 @@ func appendSystemSpecificProcessesMetrics(metrics pdata.MetricSlice, startIndex 
 	return nil
 }
 
-func initializeProcessesCountMetric(metric pdata.Metric, now pdata.TimestampUnixNano, misc *load.MiscStat) {
+func initializeProcessesCountMetric(metric pdata.Metric, now pdata.Timestamp, misc *load.MiscStat) {
 	metadata.Metrics.SystemProcessesCount.Init(metric)
 
 	ddps := metric.IntSum().DataPoints()
@@ -54,7 +54,7 @@ func initializeProcessesCountMetric(metric pdata.Metric, now pdata.TimestampUnix
 	initializeProcessesCountDataPoint(ddps.At(1), now, metadata.LabelProcessesStatus.Blocked, int64(misc.ProcsBlocked))
 }
 
-func initializeProcessesCountDataPoint(dataPoint pdata.IntDataPoint, now pdata.TimestampUnixNano, statusLabel string, value int64) {
+func initializeProcessesCountDataPoint(dataPoint pdata.IntDataPoint, now pdata.Timestamp, statusLabel string, value int64) {
 	labelsMap := dataPoint.LabelsMap()
 	labelsMap.Insert(metadata.Labels.ProcessesStatus, statusLabel)
 	dataPoint.SetTimestamp(now)