+++ translator/trace/zipkin/zipkinv2_to_traces.go
@@ -136,8 +136,8 @@ func zSpanToInternal(zspan *zipkinmodel.SpanModel, tags map[string]string, dest 
 
 	dest.SetName(zspan.Name)
 	startNano := zspan.Timestamp.UnixNano()
-	dest.SetStartTime(pdata.TimestampUnixNano(startNano))
-	dest.SetEndTime(pdata.TimestampUnixNano(startNano + zspan.Duration.Nanoseconds()))
+	dest.SetStartTime(pdata.Timestamp(startNano))
+	dest.SetEndTime(pdata.Timestamp(startNano + zspan.Duration.Nanoseconds()))
 	dest.SetKind(zipkinKindToSpanKind(zspan.Kind, tags))
 
 	populateSpanStatus(tags, dest.Status())
@@ -260,7 +260,7 @@ func populateSpanEvents(zspan *zipkinmodel.SpanModel, events pdata.SpanEventSlic
 	for ix, anno := range zspan.Annotations {
 		event := events.At(ix)
 		startNano := anno.Timestamp.UnixNano()
-		event.SetTimestamp(pdata.TimestampUnixNano(startNano))
+		event.SetTimestamp(pdata.Timestamp(startNano))
 
 		parts := strings.Split(anno.Value, "|")
 		partCnt := len(parts)